Fixed a bug in the FileComplete event handler that was causing a Type Coercion error for many users (that had debug players) and was the actual cause of the FileComplete event not firing (I thought it was the ServerData event).
Added API for updating all the updateable settings dynamically.  Removed the "UpdateUploadSettings" method.  All changes to updateable settings are now immediate.

Added a setting for indicating if the FileComplete or ServerData event should be used.  The architecture does not allow both to be called since the 'current_item' pointer will be changed by the time the second event is called.  This also required moving the addEventListener statements from queue time to upload time.  Without this change the UseServerData setting gets applied when a file is queued and any subsequent changes are ignored for already queued files.